CR feedback, clarify private vs. external-facing
